--- source/logic/seasons.js
@@ -1,279 +1,0 @@
-const { Op } = require("sequelize");
-const { DatabaseTypes } = require("../database");
-const { calculateXPMean, calculateXPStandardDeviation } = require("./shared");
-const { descendingByProperty } = require("../shared");
-const { Role, Collection } = require("discord.js");
-
-/** @type {import("../database").Database} */
-let db;
-
-/** *Sets the database pointer for the Season logic file*
- * @param {import("../database").Database} database
- */
-function setDB(database) {
-	db = database;
-}
-
-/** @param {string} companyId */
-function createSeason(companyId) {
-	return db.models.Season.create({ companyId })
-}
-
-/** @param {string} companyId */
-function findOrCreateCurrentSeason(companyId) {
-	return db.models.Season.findOrCreate({ where: { companyId, isCurrentSeason: true } });
-}
-
-/**
- * @param {string} companyId
- * @param {"current" | "previous"} type
- */
-function findOneSeason(companyId, type) {
-	switch (type) {
-		case "current":
-			return db.models.Season.findOne({ where: { companyId, isCurrentSeason: true } });
-		case "previous":
-			return db.models.Season.findOne({ where: { companyId, isPreviousSeason: true } });
-	}
-}
-
-/** *Get the number of participating bounty hunters in the specified Season*
- * @param {string} seasonId
- */
-function getParticipantCount(seasonId) {
-	return db.models.Participation.count({ where: { seasonId } });
-}
-
-/** *Counts the number of times the specified Hunter has been DQ'd*
- * @param {string} userId
- * @param {string} companyId
- */
-function getDQCount(userId, companyId) {
-	return db.models.Participation.sum("dqCount", { where: { userId, companyId } }) ?? 0;
-}
-
-/** *Returns a Map of userId to Participation for all Participations in the specified Season*
- * @param {string} seasonId
- */
-async function getParticipationMap(seasonId) {
-	/** @type {Map<string, DatabaseTypes.Participation>} */
-	const participationMap = new Map();
-	const participations = await db.models.Participation.findAll({ where: { seasonId } });
-	for (const participation of participations) {
-		participationMap.set(participation.userId, participation);
-	}
-	return participationMap;
-}
-
-/** *Get Participations of the specified Users in the specified Season*
- * @param {string} seasonId
- * @param {string[]} userIds
- */
-function bulkFindParticipations(seasonId, userIds) {
-	return db.models.Participation.findAll({ where: { seasonId, userId: { [Op.in]: userIds } } });
-}
-
-/** *Get all the Participations of a specified Hunter*
- * @param {string} userId
- * @param {string} companyId
- */
-function findHunterParticipations(userId, companyId) {
-	return db.models.Participation.findAll({ where: { userId, companyId }, order: [["createdAt", "DESC"]] });
-}
-
-/** *Find the first place Participation in the specified Company and Season*
- * @param {string} companyId
- * @param {string} seasonId
- */
-function findFirstPlaceParticipation(companyId, seasonId) {
-	return db.models.Participation.findOne({ where: { companyId, seasonId, placement: 1 } });
-}
-
-/** *Finds the Participation of the specified Season's Hunter with the most of the specified Participation property*
- * @param {string} companyId
- * @param {string} seasonId
- * @param {string} participationProperty
- */
-async function findParticipationWithTopParticipationStat(companyId, seasonId, participationProperty) {
-	const participation = await db.models.Participation.findOne({ where: { companyId, seasonId }, order: [[participationProperty, "DESC"]] });
-	if (participation === null || participation[participationProperty] === 0) {
-		return null;
-	}
-	return participation;
-}
-
-/** Calculates the XP required for the specified Hunter to reach the next Rank
- * @param {string} userId
- * @param {DatabaseTypes.Season} season
- * @param {DatabaseTypes.Rank[]} descendingRanks
- */
-async function nextRankXP(userId, season, descendingRanks) {
-	const participationMap = await getParticipationMap(season.id);
-	const participation = participationMap.get(userId);
-	if (!participation || participation.rankIndex === null || participation.rankIndex === 0) {
-		return 0;
-	}
-	const mean = calculateXPMean(participationMap);
-	const xpStandardDeviation = calculateXPStandardDeviation(participationMap, mean);
-	return Math.ceil(xpStandardDeviation * descendingRanks[participation.rankIndex - 1].threshold + mean - participation.xp);
-}
-
-
-/** Recalculate placement and rank changes based on changed XP values on Participations and updates the database
- * @param {Map<string, DatabaseTypes.Participation>} participationMap
- * @param {DatabaseTypes.Rank[]} descendingRanks
- * @param {Collection<import("discord.js").Snowflake, Role>} allGuildRoles
- */
-async function updatePlacementsAndRanks(participationMap, descendingRanks, allGuildRoles) {
-	if (participationMap.size < 1) {
-		return new Map();
-	}
-	const placementChanges = await calculatePlacementChanges(participationMap);
-	const mean = calculateXPMean(participationMap);
-	const standardDeviation = calculateXPStandardDeviation(participationMap, mean);
-
-	// Calculate Rank Changes
-	const rankChanges = {};
-	if (descendingRanks.length > 0) {
-		for (const [id, participation] of participationMap) {
-			const standardDeviationsFromMean = (participation.xp - mean) / standardDeviation;
-			let index = 0;
-			for (const rank of descendingRanks) {
-				if (standardDeviationsFromMean >= rank.threshold) {
-					break;
-				}
-				index++;
-			}
-			const previousRankIndex = participation.rankIndex === null ? descendingRanks.length : participation.rankIndex;
-			rankChanges[id] = { index: index === descendingRanks.length ? null : index, isRankUp: index < previousRankIndex };
-		}
-	}
-
-	const seasonalHunterReceipts = new Map();
-	for (const id of new Set(Object.keys(placementChanges).concat(Object.keys(rankChanges)))) {
-		const updatePayload = {};
-		const rawReceipt = {};
-		if (id in placementChanges) {
-			updatePayload.placement = placementChanges[id];
-			if (placementChanges[id] === 1) {
-				rawReceipt.topPlacement = true;
-			}
-		}
-		if (id in rankChanges) {
-			const { index, isRankUp } = rankChanges[id];
-			updatePayload.rankIndex = index;
-			if (isRankUp) {
-				const rank = descendingRanks[index];
-				rawReceipt.rankUp = { name: rank.getName(allGuildRoles, index), newRankIndex: index };
-			}
-		}
-		await participationMap.get(id).update(updatePayload);
-		seasonalHunterReceipts.set(id, rawReceipt);
-	}
-	return seasonalHunterReceipts;
-}
-
-/** *Generates a map of all of a Season's Participations' placement changes*
- * @param {Map<any, DatabaseTypes.Participation>} participationMap
- */
-async function calculatePlacementChanges(participationMap) {
-	const participationArray = Array.from(participationMap.values()).sort(descendingByProperty("xp"));
-	let recentPlacement = participationMap.size;
-	let previousScore = 0;
-	const placementChanges = {};
-	// subtract 1 to adjust for array indexes starting from 0
-	for (let i = recentPlacement - 1; i >= 0; i -= 1) {
-		const participation = participationArray[i];
-		if (participation.xp > previousScore) {
-			previousScore = participation.xp;
-			recentPlacement = i + 1;
-		}
-		if (participation.placement !== recentPlacement) {
-			placementChanges[participation.userId] = recentPlacement;
-		}
-	}
-	return placementChanges;
-}
-
-/** *Change the specified Hunter's Seasonal XP*
- * @param {string} userId
- * @param {string} companyId
- * @param {string} seasonId
- * @param {number} xp negative numbers allowed
- */
-async function changeSeasonXP(userId, companyId, seasonId, xp) {
-	const [participation, participationCreated] = await db.models.Participation.findOrCreate({ where: { userId, companyId, seasonId }, defaults: { xp } });
-	if (!participationCreated) {
-		await participation.increment({ xp });
-	}
-	return participationCreated;
-}
-
-/**
- * @param {string} companyId
- * @param {string} stat
- */
-async function incrementSeasonStat(companyId, stat) {
-	const [season] = await findOrCreateCurrentSeason(companyId);
-	return season.increment(stat);
-}
-
-/**
- * @param {string} userId
- * @param {string} companyId
- * @param {string} seasonId
- */
-async function toggleHunterSeasonDisqualification(userId, companyId, seasonId) {
-	await db.models.User.findOrCreate({ where: { id: userId } });
-	const [participation, participationCreated] = await db.models.Participation.findOrCreate({ where: { userId, companyId, seasonId }, defaults: { isRankDisqualified: true } });
-	if (!participationCreated) {
-		await participation.update({ isRankDisqualified: !participation.isRankDisqualified });
-	}
-	if (participationCreated || participation.isRankDisqualified) {
-		await participation.increment("dqCount");
-	}
-	return participation;
-}
-
-/** @param {string} companyId */
-async function deleteCompanySeasons(companyId) {
-	await db.models.Participation.destroy({ where: { companyId } });
-	return db.models.Season.destroy({ where: { companyId } });
-}
-
-/**
- * @param {string} userId
- * @param {string} companyId
- */
-async function deleteHunterParticipations(userId, companyId) {
-	return db.models.Participation.destroy({ where: { userId, companyId } });
-}
-
-/** *Delete all Participations associated with the specified Season*
- * @param {string} seasonId
- */
-function deleteSeasonParticipations(seasonId) {
-	return db.models.Participation.destroy({ where: { seasonId } });
-}
-
-module.exports = {
-	setDB,
-	createSeason,
-	findOrCreateCurrentSeason,
-	findOneSeason,
-	getParticipantCount,
-	getDQCount,
-	getParticipationMap,
-	bulkFindParticipations,
-	findHunterParticipations,
-	findFirstPlaceParticipation,
-	findParticipationWithTopParticipationStat,
-	nextRankXP,
-	updatePlacementsAndRanks,
-	changeSeasonXP,
-	incrementSeasonStat,
-	toggleHunterSeasonDisqualification,
-	deleteCompanySeasons,
-	deleteHunterParticipations,
-	deleteSeasonParticipations
-}
